+++ src/test/kotlin/snyk/oss/OssServiceTest.kt
@@ -89,12 +89,12 @@ class OssServiceTest : LightPlatformTestCase() {
         assertTrue(cliCommands.contains("--sub-project=snyk"))
     }
 
-    fun testBuildCliCommandsListWithRiderIde() {
+    fun testBuildCliCommandsListContainsAllProjects() {
         setupDummyCliFile()
         val pluginInformation = PluginInformation(
             integrationName = "INTEGRATION_NAME",
             integrationVersion = "1.0.0",
-            integrationEnvironment = "INTEGRATION_ENV_RIDER",
+            integrationEnvironment = "INTEGRATION_ENV_IJ",
             integrationEnvironmentVersion = "1.0.0"
         )
 
@@ -153,7 +153,7 @@ class OssServiceTest : LightPlatformTestCase() {
 
         every {
             mockRunner.execute(
-                listOf(getCliFile().absolutePath, "test", "--json"), project.basePath!!, project = project
+                listOf(getCliFile().absolutePath, "test", "--json", "--all-projects"), project.basePath!!, project = project
             )
         } returns """
               {
@@ -182,7 +182,7 @@ class OssServiceTest : LightPlatformTestCase() {
 
         every {
             mockRunner.execute(
-                listOf(getCliFile().absolutePath, "test", "--json"), project.basePath!!, project = project
+                listOf(getCliFile().absolutePath, "test", "--json", "--all-projects"), project.basePath!!, project = project
             )
         } returns getResourceAsString("group-vulnerabilities-test.json")
 
@@ -208,7 +208,7 @@ class OssServiceTest : LightPlatformTestCase() {
 
         every {
             mockRunner.execute(
-                listOf(getCliFile().absolutePath, "test", "--json"), project.basePath!!, project = project
+                listOf(getCliFile().absolutePath, "test", "--json", "--all-projects"), project.basePath!!, project = project
             )
         } returns getResourceAsString("licence-vulnerabilities.json")